requirements.txt
asgiref
crispy-bootstrap5  
distlib
Django
django-crispy-forms
django-environ     
django-filter      
djangorestframework
filelock
Markdown
MouseInfo
pillow
platformdirs
psycopg2
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
PyScreeze
pytweening
sqlparse
tzdata
virtualenv